pyproject.toml
[project]
name = "lifesite"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiosqlite>=0.22.1",
    "bcrypt>=5.0.0",
    "fastapi>=0.136.1",
    "httpx>=0.28.1",
    "pyjwt>=2.12.1",
    "python-dotenv>=1.2.2",
    "python-jose>=3.5.0",
    "requests>=2.33.1",
    "sqlalchemy>=2.0.49",
    "uvicorn[standard]>=0.46.0",
]
[tool.uv]
pythonpath = ["src/backend"]
